--- gsmgogo-api/src/main/java/com/example/gsmgogo/domain/user/repository/UserJpaRepository.java
@@ -1,10 +1,0 @@
-package com.example.gsmgogo.domain.user.repository;
-
-import com.example.gsmgogo.domain.user.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
-    Optional<UserEntity> findByUserEmail(String userEmail);
-}
