--- gsmgogo-api/src/main/java/com/example/gsmgogo/domain/user/repository/ㅁ.java
@@ -1,12 +1,0 @@
-package com.example.gsmgogo.domain.user.repository;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ㅁ {
-    @GetMapping("/asdf")
-    public String a() {
-        return "OK#";
-    }
-}
