--- gsmgogo-api/src/main/java/com/example/gsmgogo/global/config/SecurityConfig.java
@@ -1,42 +1,0 @@
-package com.example.gsmgogo.global.config;
-
-import com.example.gsmgogo.global.security.jwt.JwtTokenProvider;
-import com.example.gsmgogo.global.security.jwt.filter.JwtTokenFilter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@EnableWebSecurity
-@Configuration
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final JwtTokenProvider jwtTokenProvider;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-
-        http.csrf(AbstractHttpConfigurer::disable);
-
-        http.sessionManagement((sessionManagement) ->
-                sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-        );
-
-        http.authorizeHttpRequests((authorizeHttpRequests) ->
-                authorizeHttpRequests
-                        .anyRequest().permitAll()
-        );
-
-        http.addFilterBefore(new JwtTokenFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class);
-
-        http.cors(AbstractHttpConfigurer::disable);
-
-        return http.build();
-    }
-}
