--- gsmgogo-api/src/main/java/com/example/gsmgogo/global/exception/model/ExceptionResponseEntity.java
@@ -1,8 +1,0 @@
-package com.example.gsmgogo.global.exception.model;
-
-public record ExceptionResponseEntity(String message) {
-    public static ExceptionResponseEntity of(final Throwable exception) {
-        return new ExceptionResponseEntity(exception.getMessage());
-    }
-
-}
