--- gsmgogo-api/src/main/java/com/example/gsmgogo/global/security/jwt/dto/TokenResponse.java
@@ -1,13 +1,0 @@
-package com.example.gsmgogo.global.security.jwt.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class TokenResponse {
-    private final String accessToken;
-    private final String refreshToken;
-}
