--- gsmgogo-api/src/main/java/com/example/gsmgogo/global/security/principle/AuthDetailsService.java
@@ -1,26 +1,0 @@
-package com.example.gsmgogo.global.security.principle;
-
-import com.example.gsmgogo.domain.user.repository.UserJpaRepository;
-import com.example.gsmgogo.global.exception.error.ExpectedException;
-import com.example.gsmgogo.domain.user.entity.UserEntity;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AuthDetailsService implements UserDetailsService {
-
-    private final UserJpaRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
-        UserEntity user = userRepository.findByUserEmail(userEmail)
-                .orElseThrow(() -> new ExpectedException("유저를 찾을 수 없습니다.", HttpStatus.NOT_FOUND));
-
-        return new AuthDetails(user);
-    }
-}
