+++ gsmgogo-api/src/main/java/team/gsmgogo/GsmgogoApiApplication.java
@@ -1,0 +1,13 @@
+package team.gsmgogo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GsmgogoApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GsmgogoApiApplication.class, args);
+    }
+
+}
