+++ gsmgogo-api/src/main/java/team/gsmgogo/domain/user/repository/UserJpaRepository.java
@@ -1,0 +1,10 @@
+package team.gsmgogo.domain.user.repository;
+
+import team.gsmgogo.domain.user.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
+    Optional<UserEntity> findByUserEmail(String userEmail);
+}
