+++ gsmgogo-api/src/main/java/team/gsmgogo/global/exception/model/ExceptionResponseEntity.java
@@ -1,0 +1,8 @@
+package team.gsmgogo.global.exception.model;
+
+public record ExceptionResponseEntity(String message) {
+    public static ExceptionResponseEntity of(final Throwable exception) {
+        return new ExceptionResponseEntity(exception.getMessage());
+    }
+
+}
