+++ gsmgogo-api/src/main/java/team/gsmgogo/global/security/jwt/dto/TokenResponse.java
@@ -1,0 +1,13 @@
+package team.gsmgogo.global.security.jwt.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class TokenResponse {
+    private final String accessToken;
+    private final String refreshToken;
+}
