+++ gsmgogo-api/src/main/java/team/gsmgogo/global/security/principle/AuthDetails.java
@@ -1,0 +1,61 @@
+package team.gsmgogo.global.security.principle;
+
+import team.gsmgogo.domain.user.entity.UserEntity;
+import team.gsmgogo.domain.user.enums.Role;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.ArrayList;
+import java.util.Collection;
+
+@RequiredArgsConstructor
+@Getter
+public class AuthDetails implements UserDetails {
+
+    private final UserEntity user;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        Role role = user.getRole();
+        String authority = role.getRole();
+
+        SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(simpleGrantedAuthority);
+
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
+        return null;
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getUserEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
