+++ gsmgogo-api/src/main/java/team/gsmgogo/global/security/principle/AuthDetailsService.java
@@ -1,0 +1,26 @@
+package team.gsmgogo.global.security.principle;
+
+import team.gsmgogo.domain.user.repository.UserJpaRepository;
+import team.gsmgogo.global.exception.error.ExpectedException;
+import team.gsmgogo.domain.user.entity.UserEntity;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AuthDetailsService implements UserDetailsService {
+
+    private final UserJpaRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
+        UserEntity user = userRepository.findByUserEmail(userEmail)
+                .orElseThrow(() -> new ExpectedException("유저를 찾을 수 없습니다.", HttpStatus.NOT_FOUND));
+
+        return new AuthDetails(user);
+    }
+}
