--- gsmgogo-api/src/test/java/com/example/gsmgogo/GsmgogoApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.gsmgogo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GsmgogoApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
