+++ gsmgogo-api/src/test/java/team/gsmgogo/GsmgogoApiApplicationTests.java
@@ -1,0 +1,13 @@
+package team.gsmgogo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GsmgogoApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
