--- gsmgogo-batch/src/main/java/com/example/gsmgogo/GsmgogoBatchApplication.java
@@ -1,13 +1,0 @@
-package com.example.gsmgogo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GsmgogoBatchApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GsmgogoBatchApplication.class, args);
-    }
-
-}
