--- gsmgogo-batch/src/main/java/com/example/gsmgogo/ㅁ.java
@@ -1,12 +1,0 @@
-package com.example.gsmgogo;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ㅁ {
-    @GetMapping("/asd")
-    public String a() {
-        return "OK#";
-    }
-}
