+++ gsmgogo-batch/src/main/java/team/gsmgogo/GsmgogoBatchApplication.java
@@ -1,0 +1,13 @@
+package team.gsmgogo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GsmgogoBatchApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GsmgogoBatchApplication.class, args);
+    }
+
+}
