--- gsmgogo-batch/src/main/resources/application.yml
@@ -1,18 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:h2:mem:gogodb
-    driver-class-name: org.h2.Driver
-    username: sa
-    password:
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
-  jpa:
-    hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-        format_sql: true
-        use_sql_comments: true
-    database-platform: org.hibernate.dialect.H2Dialect
