--- gsmgogo-batch/src/test/java/com/example/gsmgogo/GsmgogoBatchApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.gsmgogo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GsmgogoBatchApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
