+++ gsmgogo-batch/src/test/java/team/gsmgogo/GsmgogoBatchApplicationTests.java
@@ -1,0 +1,13 @@
+package team.gsmgogo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GsmgogoBatchApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
