--- gsmgogo-entity/src/main/java/com/example/gsmgogo/domain/user/enums/GradeEnum.java
@@ -1,15 +1,0 @@
-package com.example.gsmgogo.domain.user.enums;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum GradeEnum {
-
-    ONE("ONE"),
-    TWO("TWO"),
-    THREE("THREE");
-
-    private final String role;
-}
