--- gsmgogo-entity/src/main/java/com/example/gsmgogo/domain/user/enums/Role.java
@@ -1,15 +1,0 @@
-package com.example.gsmgogo.domain.user.enums;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum Role {
-
-    USER("USER"),
-    PLAYER("PLAYER"),
-    LEADER("LEADER");
-
-    private final String role;
-}
