+++ gsmgogo-entity/src/main/java/team/gsmgogo/domain/user/enums/GradeEnum.java
@@ -1,0 +1,15 @@
+package team.gsmgogo.domain.user.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum GradeEnum {
+
+    ONE("ONE"),
+    TWO("TWO"),
+    THREE("THREE");
+
+    private final String role;
+}
