+++ gsmgogo-entity/src/main/java/team/gsmgogo/domain/user/enums/Role.java
@@ -1,0 +1,15 @@
+package team.gsmgogo.domain.user.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum Role {
+
+    USER("USER"),
+    PLAYER("PLAYER"),
+    LEADER("LEADER");
+
+    private final String role;
+}
