+++ gsmgogo-entity/src/main/java/team/gsmgogo/domain/user/enums/SchoolRole.java
@@ -1,0 +1,14 @@
+package team.gsmgogo.domain.user.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum SchoolRole {
+
+    STUDENT("STUDENT"),
+    TEACHER("TEACHER");
+
+    private final String role;
+}
